+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 1.11
 ### Capacités
+* Animer un cadavre, rang 2 de la voie de la magie maléfique.
 * Agripper, rang 2 de la voie des créature volantes
 * Destruction des mort-vivants peut maintenant utiliser les options de tempête de mana
 * Mur de force utilise maintenant une image par défaut.